Add files via upload
Learned how to remove and replace items in our HashMap with all of its details. One interesting thing is that you can use any object for the key and value in a HashMap. In other words, it doesn't solely have to be a primitive data types or Strings.

This is also an extension of MapProgram (hence the 2). The reason I did this was because there were too many comments in the previous one, but I wanted to add more.